Added delete auth

+++ backend/controllers/emails.js
@@ -71,6 +71,11 @@ const deleteEmails = (req, res) => {
       message: "You must supply the id for the email you want to delete.",
       error: true,
     });
+  let Email = EmailsSchema.find({ id: req.params.id });
+  if(Email.addedBy !== req.user.id) {
+    res.status(400).json({ message: 'You are not authorized for that', error: true });
+    return; // end the process
+  }
   // do some verification here -- are they allowed to delete this?
   EmailsSchema.findOneAndDelete({ id: req.params.id }).exec(function (err) {
     if (err)
